[IMP] shopinvader_anonymous_partner: Add a cookie helper
