+++ shopinvader_anonymous_partner/README.rst
@@ -7,7 +7,7 @@ Shopinvader Anonymous Partner
    !! This file is generated by oca-gen-addon-readme !!
    !! changes will be overwritten.                   !!
    !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
-   !! source digest: sha256:7b6ee27b03a4ec3fe64444468334b4a0cb81d53ecb73f7b8a07bcddfddf968f5
+   !! source digest: sha256:025b5acbccc95cad32a4b1acae176d3d99aa99ecd0bbbf2411f8c6ac3f7f4f03
    !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
 
 .. |badge1| image:: https://img.shields.io/badge/maturity-Beta-yellow.png
@@ -33,7 +33,7 @@ a cookie to establish a session for these partners.
 Usage
 =====
 
-The following model methods are added to ``res.partner``:
+The following model methods are added in a helper model:
 
 ``_create_anonymous_partner__cookie(response)``
