+++ shopinvader_anonymous_partner/models/__init__.py
@@ -1,1 +1,2 @@
 from . import res_partner
+from . import cookie_helper
