+++ shopinvader_fastapi_auth_jwt/dependencies.py
@@ -29,9 +29,8 @@ def auth_jwt_authenticated_or_anonymous_partner(
 ) -> Partner:
     if auth_jwt_partner:
         return auth_jwt_partner
-    anonymous_partner = env["res.partner"]._get_anonymous_partner__cookie(
-        request.cookies
-    )
+    cookie_helper = env["shopinvader_anonymous_partner.cookie.helper"]
+    anonymous_partner = cookie_helper._get_anonymous_partner__cookie(request.cookies)
     if anonymous_partner:
         return env["res.partner"].browse(anonymous_partner.id)
     _logger.info("JWT authentication failed and no anonymous partner cookie found.")
@@ -53,9 +52,8 @@ def auth_jwt_authenticated_or_anonymous_partner_autocreate(
 ) -> Partner:
     if auth_jwt_partner:
         return auth_jwt_partner
-    anonymous_partner = env["res.partner"]._get_anonymous_partner__cookie(
-        request.cookies
-    )
+    cookie_helper = env["shopinvader_anonymous_partner.cookie.helper"]
+    anonymous_partner = cookie_helper._get_anonymous_partner__cookie(request.cookies)
     if not anonymous_partner:
         if payload:
             _logger.info(
@@ -63,7 +61,5 @@ def auth_jwt_authenticated_or_anonymous_partner_autocreate(
                 "Not attempting to create an anonymous partner."
             )
             raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED)
-        anonymous_partner = env["res.partner"]._create_anonymous_partner__cookie(
-            response
-        )
+        anonymous_partner = cookie_helper._create_anonymous_partner__cookie(response)
     return env["res.partner"].browse(anonymous_partner.id)
